refactor: Improve folder structure and and use design patterns
