+++ main.py
@@ -1,19 +1,20 @@
 from fastapi import FastAPI
-from routers import router
+from routes.routers import router
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel
 from sklearn.feature_extraction.text import CountVectorizer
-from sklearn.naive_bayes import MultinomialNB
 import uvicorn
-import numpy as np
 from os import environ
 
 
 app = FastAPI()
-
-
-# Aquí puedes agregar configuraciones adicionales, como CORS
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
 app.include_router(router)
